Add accessor returning layer ID and size
When an image is pulled down from the repository, as each layer begins
to be retrieved, the ImageBuildChunk object will contain its eventual
compressed size.  This new ImageBuildChunk::image_layer_bytes() function
will retrieve the corresponding layer ID and size for each chunk in
which it is available.

This information can be used to calculate the eventual total image size,
before the entire image is downloaded.  This is useful in constrained
environments that must place a limit on image size, so that an image
download may be aborted when it is clear that it is exceeding safe
limits.

image_layer_bytes() solves a deficiency in
ImageBuildChunk::total_image_bytes(), which doesn't allow the layer ID's
to be tracked, thus making its return value often useless as chunks for
several layers may be coming in interleaved.  That function should
probably be deprecated.

The examples/imagepull_layers.rs file provides a model for usage.